Change lang.nb to lang.no
