+++ spacy/lang/no/examples.py
