+++ spacy/lang/no/lemmatizer/LICENSE
