+++ spacy/lang/no/lemmatizer/lemma_lookup.json
