+++ spacy/lang/no/lemmatizer/lemma_rules.json
