+++ spacy/lang/no/morph_rules.py
