+++ spacy/lang/no/punctuation.py
