+++ spacy/lang/no/stop_words.py
