+++ spacy/lang/no/syntax_iterators.py
