+++ spacy/lang/no/tag_map.py
