+++ spacy/tests/lang/no/test_tokenizer.py
@@ -2,6 +2,7 @@
 from __future__ import unicode_literals
 
 import pytest
+from spacy.util import get_lang_class
 
 
 NB_TOKEN_EXCEPTION_TESTS = [
@@ -17,7 +18,12 @@ NB_TOKEN_EXCEPTION_TESTS = [
 
 
 @pytest.mark.parametrize("text,expected_tokens", NB_TOKEN_EXCEPTION_TESTS)
-def test_nb_tokenizer_handles_exception_cases(nb_tokenizer, text, expected_tokens):
-    tokens = nb_tokenizer(text)
+def test_no_tokenizer_handles_exception_cases(no_tokenizer, text, expected_tokens):
+    tokens = no_tokenizer(text)
     token_list = [token.text for token in tokens if not token.is_space]
     assert expected_tokens == token_list
+
+
+def test_no_nb_warning():
+    with pytest.warns(UserWarning):
+        get_lang_class("nb")
