+++ spacy/tests/lang/test_initialize.py
@@ -10,7 +10,7 @@ from spacy.util import get_lang_class
 # excluded: ja, ru, th, uk, vi, zh
 LANGUAGES = ["af", "ar", "bg", "bn", "ca", "cs", "da", "de", "el", "en", "es",
              "et", "fa", "fi", "fr", "ga", "he", "hi", "hr", "hu", "id", "is",
-             "it", "kn", "lt", "lv", "nb", "nl", "pl", "pt", "ro", "si", "sk",
+             "it", "kn", "lt", "lv", "no", "nl", "pl", "pt", "ro", "si", "sk",
              "sl", "sq", "sr", "sv", "ta", "te", "tl", "tr", "tt", "ur"]
 # fmt: on
